refactor: extract internal/device package from internal/sensors
Moves Device model, DeviceStore interface, postgres implementation, and
DeviceService into internal/device. Sensors package keeps thin wrappers
for backward compat. Platform middleware now uses device.Info/ContextKey.

Closes #97
