+++ internal/sensors/model.go
@@ -1,9 +1,10 @@
 package sensors
 
 import (
-	"context"
 	"errors"
 	"time"
+
+	"github.com/fishhub-oss/fishhub-server/internal/device"
 )
 
 type Peripheral struct {
@@ -28,7 +29,7 @@ type ScheduleWindow struct {
 }
 
 var ErrNotAnActuator           = errors.New("peripheral is not an actuator")
-var ErrDeviceNotFound          = errors.New("device not found")
+var ErrDeviceNotFound          = device.ErrNotFound
 var ErrCodeNotFound            = errors.New("provisioning code not found")
 var ErrCodeAlreadyUsed         = errors.New("provisioning code already used")
 var ErrInvalidCommand          = errors.New("action must be 'set' or 'schedule'")
@@ -36,17 +37,3 @@ var ErrInfluxWrite             = errors.New("failed to persist reading")
 var ErrPeripheralNotFound      = errors.New("peripheral not found")
 var ErrPeripheralAlreadyExists = errors.New("peripheral already exists")
 var ErrPeripheralPinInUse      = errors.New("peripheral pin already in use")
-
-type DeviceInfo struct {
-	DeviceID string
-	UserID   string
-}
-
-type contextKey string
-
-const DeviceContextKey contextKey = "device"
-
-func DeviceFromContext(ctx context.Context) (DeviceInfo, bool) {
-	info, ok := ctx.Value(DeviceContextKey).(DeviceInfo)
-	return info, ok
-}
