+++ internal/sensors/store.go
@@ -3,42 +3,12 @@ package sensors
 import (
 	"context"
 	"database/sql"
-	"time"
+
+	"github.com/fishhub-oss/fishhub-server/internal/device"
 )
 
-type Device struct {
-	ID        string
-	UserID    string
-	Name      string
-	CreatedAt time.Time
-}
-
-// ActivationStatus holds the device's MQTT readiness state.
-type ActivationStatus struct {
-	Ready        bool
-	MQTTUsername string
-	MQTTPassword string
-	MQTTHost     string
-	MQTTPort     int
-}
-
-type DeviceStore interface {
-	ListByUserID(ctx context.Context, userID string) ([]Device, error)
-	// FindByID looks up a device by its ID regardless of owner.
-	// Returns ErrDeviceNotFound if the device does not exist or is soft-deleted.
-	FindByID(ctx context.Context, deviceID string) (Device, error)
-	FindByIDAndUserID(ctx context.Context, deviceID, userID string) (Device, error)
-	// PatchDevice updates the name of the device owned by userID.
-	// Returns ErrDeviceNotFound if the device does not exist or is not owned by the user.
-	PatchDevice(ctx context.Context, deviceID, userID, name string) (Device, error)
-	// DeleteDevice soft-deletes the device and returns its mqtt_username for cleanup.
-	// Returns ErrDeviceNotFound if the device does not exist or is not owned by the user.
-	DeleteDevice(ctx context.Context, deviceID, userID string) (mqttUsername string, err error)
-	// GetActivationStatus returns whether the device's MQTT credentials are ready.
-	// Ready = credentials present in DB AND no pending/processing outbox event for the device.
-	// Returns ErrDeviceNotFound if the device does not exist.
-	GetActivationStatus(ctx context.Context, deviceID string) (ActivationStatus, error)
-}
+// DeviceStore is an alias for device.Store kept for backward compatibility within this package.
+type DeviceStore = device.Store
 
 type PeripheralStore interface {
 	// CreatePeripheral inserts a new peripheral for the device owned by userID.
